Add files via upload
This was the starter file for SQL and Oracle relational database component in a Masters level CS/IT subject completed at the Swinburne University of Technology in 2018. Was unable to retrieve the submitted portfolio, but these tasks were all marked by faculty members to be granted a Distinction subject mark.